Relaxed semantic of compiler's "target" attribute
Before this change the semantic to check if a compiler could be viable
for a given target was exact match. This made sense as the finest
granularity of targets was architecture families.

As now we can target micro-architectures, this commit changes the
semantic by interpreting as the architecture family what is stored in
the compiler's "target" attribute. A compiler is then a viable choice
if the target being concretized belongs to the same family.

Similarly when a new compiler is detected the architecture family is
stored in the "target" attribute.