Feature aliases are now defined in targets.json
MicroArchitecture objects are now also responsible for checks on feature
aliases.

Feature aliases are defined in targets.json and map a name (the "alias")
to a list of rules that must be met for the test to be successful. The
rules that are available can be extended later using a decorator.